+++ PolySport/Data/Migrations/20260903183625_AddGoalkeeperFlag.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace PolySport.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddGoalkeeperFlag : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "IsGoalkeeper",
+                table: "MatchPlayers",
+                type: "bit",
+                nullable: false,
+                defaultValue: false);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "IsGoalkeeper",
+                table: "MatchPlayers");
+        }
+    }
+}